refactor: improve SRP in WOFF2 decoder and Flex layout
- Modularized `src/fonts/woff2/decoder.ts` into `utils.ts`, `sfnt-builder.ts`, `glyf-reconstructor.ts`, and `hmtx-reconstructor.ts`.
- Modularized `src/layout/strategies/flex.ts` into a new `src/layout/strategies/flex/` directory with `types.ts`, `utils.ts`, `line-builder.ts`, `distributor.ts`, and `alignment.ts`.
- Ensured all tests pass and addressed code review feedback regarding bridge functions and robust auto checks.
