+++ src/layout/strategies/flex.ts
@@ -3,257 +3,26 @@ import { AlignItems, Display, JustifyContent } from "../../css/enums.js";
 import type { LayoutContext, LayoutStrategy } from "../pipeline/strategy.js";
 import { adjustForBoxSizing, containingBlock, resolveBoxMetrics, resolveWidthBlock } from "../utils/node-math.js";
 import { resolveLength, isAutoLength } from "../../css/length.js";
-import type { LengthLike } from "../../css/length.js";
-import type { FlexDirection, AlignSelfValue } from "../../css/style.js";
 import { GapCalculator, calculateTotalGap } from "../utils/gap-calculator.js";
-
-function blockifyFlexItemDisplay(display: Display): Display {
-  switch (display) {
-    case Display.Inline:
-    case Display.InlineBlock:
-      return Display.Block;
-    case Display.InlineFlex:
-      return Display.Flex;
-    case Display.InlineGrid:
-      return Display.Grid;
-    case Display.InlineTable:
-      return Display.Table;
-    default:
-      return display;
-  }
-}
-
-function allowsPreferredShrink(originalDisplay: Display, effectiveDisplay: Display): boolean {
-  switch (effectiveDisplay) {
-    case Display.Flex:
-    case Display.InlineFlex:
-    case Display.Grid:
-    case Display.InlineGrid:
-      return false;
-    default:
-      break;
-  }
-  if (effectiveDisplay === originalDisplay) {
-    return true;
-  }
-  return originalDisplay === Display.Inline;
-}
-
-interface FlexItemMetrics {
-  node: LayoutNode;
-  originalDisplay: Display;
-  effectiveDisplay: Display;
-  mainMarginStart: number;
-  mainMarginEnd: number;
-  crossMarginStart: number;
-  crossMarginEnd: number;
-  mainSize: number;
-  crossSize: number;
-  mainContribution: number;
-  crossContribution: number;
-  flexGrow: number;
-  flexShrink: number;
-}
-
-interface FlexLine {
-  items: FlexItemMetrics[];
-  // Includes item contributions and explicit main-axis gaps between items.
-  mainSizeWithGaps: number;
-  // Max cross contribution among line items.
-  crossSize: number;
-}
-
-function buildFlexLines(items: FlexItemMetrics[], containerMainSize: number, mainAxisGap: number): FlexLine[] {
-  if (items.length === 0) {
-    return [];
-  }
-
-  const lines: FlexLine[] = [];
-  let current: FlexLine = { items: [], mainSizeWithGaps: 0, crossSize: 0 };
-
-  for (const item of items) {
-    const addition =
-      current.items.length === 0 ? item.mainContribution : mainAxisGap + item.mainContribution;
-    const shouldWrap =
-      current.items.length > 0 &&
-      containerMainSize > 0 &&
-      current.mainSizeWithGaps + addition > containerMainSize + 0.01;
-
-    if (shouldWrap) {
-      lines.push(current);
-      current = { items: [], mainSizeWithGaps: 0, crossSize: 0 };
-    }
-
-    if (current.items.length > 0) {
-      current.mainSizeWithGaps += mainAxisGap;
-    }
-    current.items.push(item);
-    current.mainSizeWithGaps += item.mainContribution;
-    current.crossSize = Math.max(current.crossSize, item.crossContribution);
-  }
-
-  if (current.items.length > 0) {
-    lines.push(current);
-  }
-
-  return lines;
-}
-
-function calculateLinesCrossSize(lines: FlexLine[], crossAxisGap: number): number {
-  if (lines.length === 0) {
-    return 0;
-  }
-  const totalLines = lines.reduce((sum, line) => sum + line.crossSize, 0);
-  return totalLines + calculateTotalGap(crossAxisGap, lines.length);
-}
-
-function refreshFlexItemSizes(item: FlexItemMetrics, isRow: boolean): void {
-  item.mainSize = isRow ? item.node.box.borderBoxWidth : item.node.box.borderBoxHeight;
-  item.crossSize = isRow ? item.node.box.borderBoxHeight : item.node.box.borderBoxWidth;
-  item.mainContribution = item.mainSize + item.mainMarginStart + item.mainMarginEnd;
-  item.crossContribution = item.crossSize + item.crossMarginStart + item.crossMarginEnd;
-}
-
-function recomputeLineMetrics(line: FlexLine, mainAxisGap: number): void {
-  let mainSizeWithGaps = 0;
-  let crossSize = 0;
-  for (let i = 0; i < line.items.length; i++) {
-    const item = line.items[i];
-    if (i > 0) {
-      mainSizeWithGaps += mainAxisGap;
-    }
-    mainSizeWithGaps += item.mainContribution;
-    crossSize = Math.max(crossSize, item.crossContribution);
-  }
-  line.mainSizeWithGaps = mainSizeWithGaps;
-  line.crossSize = crossSize;
-}
-
-function relayoutFlexItemForMainSize(
-  container: LayoutNode,
-  item: FlexItemMetrics,
-  context: LayoutContext,
-  targetMainSize: number,
-  isRow: boolean,
-): void {
-  if (!Number.isFinite(targetMainSize) || targetMainSize < 0) {
-    return;
-  }
-
-  const prevContainerWidth = container.box.contentWidth;
-  const prevContainerHeight = container.box.contentHeight;
-  const targetContribution = targetMainSize + item.mainMarginStart + item.mainMarginEnd;
-
-  let displayMutated = false;
-  if (item.node.style.display !== item.effectiveDisplay) {
-    item.node.style.display = item.effectiveDisplay;
-    displayMutated = true;
-  }
-
-  try {
-    if (isRow) {
-      container.box.contentWidth = Math.max(0, targetContribution);
-    } else {
-      container.box.contentHeight = Math.max(0, targetContribution);
-    }
-    context.layoutChild(item.node);
-  } finally {
-    container.box.contentWidth = prevContainerWidth;
-    container.box.contentHeight = prevContainerHeight;
-    if (displayMutated) {
-      item.node.style.display = item.originalDisplay;
-    }
-  }
-}
-
-function distributeFlexGrowAcrossLines(
-  lines: FlexLine[],
-  container: LayoutNode,
-  context: LayoutContext,
-  containerMainSize: number,
-  mainAxisGap: number,
-  isRow: boolean,
-): void {
-  if (lines.length === 0 || !(containerMainSize > 0)) {
-    return;
-  }
-
-  for (const line of lines) {
-    const freeSpace = containerMainSize - line.mainSizeWithGaps;
-    if (!(freeSpace > 0)) {
-      continue;
-    }
-    const growItems = line.items.filter((item) => item.flexGrow > 0);
-    const totalGrow = growItems.reduce((sum, item) => sum + item.flexGrow, 0);
-    if (!(totalGrow > 0)) {
-      continue;
-    }
-
-    for (const item of growItems) {
-      const delta = (freeSpace * item.flexGrow) / totalGrow;
-      const targetMainSize = Math.max(0, item.mainSize + delta);
-      if (Math.abs(targetMainSize - item.mainSize) < 0.01) {
-        continue;
-      }
-      relayoutFlexItemForMainSize(container, item, context, targetMainSize, isRow);
-      refreshFlexItemSizes(item, isRow);
-    }
-
-    recomputeLineMetrics(line, mainAxisGap);
-  }
-}
-
-interface AlignContentResolution {
-  lineCrossSizes: number[];
-  initialOffset: number;
-  additionalGap: number;
-}
-
-function resolveAlignContentLayout(
-  lines: FlexLine[],
-  alignContent: string,
-  containerCrossSize: number,
-  crossAxisGap: number,
-): AlignContentResolution {
-  const lineCrossSizes = lines.map((line) => line.crossSize);
-  if (lines.length === 0) {
-    return { lineCrossSizes, initialOffset: 0, additionalGap: 0 };
-  }
-
-  const naturalCross = calculateLinesCrossSize(lines, crossAxisGap);
-  const freeSpace = Math.max(0, containerCrossSize - naturalCross);
-
-  switch (alignContent) {
-    case "stretch":
-      if (freeSpace > 0) {
-        const extraPerLine = freeSpace / lines.length;
-        for (let i = 0; i < lineCrossSizes.length; i++) {
-          lineCrossSizes[i] += extraPerLine;
-        }
-      }
-      return { lineCrossSizes, initialOffset: 0, additionalGap: 0 };
-    case "flex-end":
-      return { lineCrossSizes, initialOffset: freeSpace, additionalGap: 0 };
-    case "center":
-      return { lineCrossSizes, initialOffset: freeSpace / 2, additionalGap: 0 };
-    case "space-between":
-      if (lines.length <= 1) {
-        return { lineCrossSizes, initialOffset: 0, additionalGap: 0 };
-      }
-      return { lineCrossSizes, initialOffset: 0, additionalGap: freeSpace / (lines.length - 1) };
-    case "space-around": {
-      const gap = freeSpace / lines.length;
-      return { lineCrossSizes, initialOffset: gap / 2, additionalGap: gap };
-    }
-    case "space-evenly": {
-      const gap = freeSpace / (lines.length + 1);
-      return { lineCrossSizes, initialOffset: gap, additionalGap: gap };
-    }
-    case "flex-start":
-    default:
-      return { lineCrossSizes, initialOffset: 0, additionalGap: 0 };
-  }
-}
+import type { FlexItemMetrics, FlexLine } from "./flex/types.js";
+import {
+  blockifyFlexItemDisplay,
+  allowsPreferredShrink,
+  isRowDirection,
+  isAutoMainSize,
+  computePreferredInlineWidth,
+  offsetLayoutSubtree,
+  resolveInitialDimension,
+  resolveFlexSize
+} from "./flex/utils.js";
+import { buildFlexLines, calculateLinesCrossSize } from "./flex/line-builder.js";
+import { distributeFlexGrowAcrossLines } from "./flex/distributor.js";
+import {
+  resolveAlignContentLayout,
+  resolveJustifySpacing,
+  computeCrossOffset,
+  resolveItemAlignment
+} from "./flex/alignment.js";
 
 export class FlexLayoutStrategy implements LayoutStrategy {
   private readonly supportedDisplays = new Set<Display>([Display.Flex, Display.InlineFlex]);
@@ -632,163 +401,4 @@ export class FlexLayoutStrategy implements LayoutStrategy {
     node.box.scrollWidth = node.box.contentWidth;
     node.box.scrollHeight = node.box.contentHeight;
   }
-}
-
-function resolveInitialDimension(specified: number | undefined, fallback: number): number {
-  if (specified !== undefined && Number.isFinite(specified)) {
-    return Math.max(specified, 0);
-  }
-  if (Number.isFinite(fallback)) {
-    return Math.max(fallback, 0);
-  }
-  return 0;
-}
-
-function resolveFlexSize(
-  value: LengthLike | undefined,
-  reference: number,
-  containerWidth: number = reference,
-  containerHeight: number = reference,
-): number | undefined {
-  if (value === undefined) {
-    return undefined;
-  }
-  if (typeof value === "number") {
-    return value;
-  }
-  if (value === "auto" || isAutoLength(value)) {
-    return undefined;
-  }
-  return resolveLength(value, reference, { auto: "reference", containerWidth, containerHeight });
-}
-
-function resolveItemAlignment(alignSelf: AlignSelfValue | undefined, containerAlign: AlignItems): AlignItems {
-  if (alignSelf && alignSelf !== "auto") {
-    return alignSelf;
-  }
-  return containerAlign;
-}
-
-function computeCrossOffset(
-  alignment: AlignItems,
-  containerSize: number,
-  itemSize: number,
-  marginStart: number,
-  marginEnd: number,
-): number {
-  const total = itemSize + marginStart + marginEnd;
-  const reference = Number.isFinite(containerSize) ? containerSize : total;
-  const freeSpace = reference - total;
-  if (freeSpace <= 0) {
-    return 0;
-  }
-  switch (alignment) {
-    case AlignItems.Center:
-      return freeSpace / 2;
-    case AlignItems.FlexEnd:
-      return freeSpace;
-    default:
-      return 0;
-  }
-}
-
-function resolveJustifySpacing(
-  justify: JustifyContent,
-  freeSpace: number,
-  itemCount: number,
-): { offset: number; gap: number } {
-  if (itemCount <= 0) {
-    return { offset: 0, gap: 0 };
-  }
-  const clamped = Number.isFinite(freeSpace) ? Math.max(freeSpace, 0) : 0;
-  switch (justify) {
-    case JustifyContent.Center:
-      return { offset: clamped / 2, gap: 0 };
-    case JustifyContent.FlexEnd:
-    case JustifyContent.End:
-    case JustifyContent.Right:
-      return { offset: clamped, gap: 0 };
-    case JustifyContent.SpaceBetween:
-      if (itemCount === 1) {
-        return { offset: 0, gap: 0 };
-      }
-      return { offset: 0, gap: clamped / (itemCount - 1) };
-    case JustifyContent.SpaceAround: {
-      const gap = clamped / itemCount;
-      return { offset: gap / 2, gap };
-    }
-    case JustifyContent.SpaceEvenly: {
-      const gap = clamped / (itemCount + 1);
-      return { offset: gap, gap };
-    }
-    default:
-      return { offset: 0, gap: 0 };
-  }
-}
-
-function isRowDirection(direction: FlexDirection): boolean {
-  return direction === "row" || direction === "row-reverse";
-}
-
-function isAutoMainSize(value: LengthLike | undefined): boolean {
-  if (value === undefined) {
-    return true;
-  }
-  if (typeof value === "number") {
-    return false;
-  }
-  if (value === "auto") {
-    return true;
-  }
-  return isAutoLength(value);
-}
-
-function computePreferredInlineWidth(node: LayoutNode): number | undefined {
-  let maxWidth = 0;
-  node.walk((desc) => {
-    if (desc.inlineRuns && desc.inlineRuns.length > 0) {
-      const localMax = desc.inlineRuns.reduce(
-        (max, run) => Math.max(max, run.lineWidth ?? run.width),
-        0,
-      );
-      if (localMax > maxWidth) {
-        maxWidth = localMax;
-      }
-      return;
-    }
-    if (!desc.lineBoxes || desc.lineBoxes.length === 0) {
-      return;
-    }
-    for (const line of desc.lineBoxes) {
-      if (!line) {
-        continue;
-      }
-      const width = typeof line.width === "number" ? line.width : 0;
-      if (width > maxWidth) {
-        maxWidth = width;
-      }
-    }
-  });
-  return maxWidth > 0 ? maxWidth : undefined;
-}
-
-function offsetLayoutSubtree(node: LayoutNode, deltaX: number, deltaY: number): void {
-  if (deltaX === 0 && deltaY === 0) {
-    return;
-  }
-  node.walk((desc) => {
-    if (desc !== node) {
-      desc.box.x += deltaX;
-      desc.box.y += deltaY;
-    }
-    desc.box.baseline += deltaY;
-
-    // Update inline runs if they exist
-    if (desc.inlineRuns && desc.inlineRuns.length > 0) {
-      for (const run of desc.inlineRuns) {
-        run.startX += deltaX;
-        run.baseline += deltaY;
-      }
-    }
-  });
 }
